Remove obj-identify
