+++ src/Makefile.src
@@ -123,7 +123,6 @@ ANGFILES = \
 	obj-chest.o \
 	obj-desc.o \
 	obj-gear.o \
-	obj-identify.o \
 	obj-ignore.o \
 	obj-info.o \
 	obj-knowledge.o \
