+++ src/game-world.c
@@ -26,7 +26,6 @@
 #include "mon-util.h"
 #include "obj-desc.h"
 #include "obj-gear.h"
-#include "obj-identify.h"
 #include "obj-knowledge.h"
 #include "obj-tval.h"
 #include "obj-util.h"
