+++ src/load.c
@@ -28,7 +28,6 @@
 #include "mon-util.h"
 #include "monster.h"
 #include "obj-gear.h"
-#include "obj-identify.h"
 #include "obj-ignore.h"
 #include "obj-make.h"
 #include "obj-pile.h"
