+++ src/main-stats.c
@@ -27,7 +27,6 @@
 #include "mon-make.h"
 #include "monster.h"
 #include "obj-gear.h"
-#include "obj-identify.h"
 #include "obj-power.h"
 #include "obj-randart.h"
 #include "obj-tval.h"
