+++ src/mon-lore.c
@@ -25,7 +25,6 @@
 #include "mon-spell.h"
 #include "mon-util.h"
 #include "obj-gear.h"
-#include "obj-identify.h"
 #include "obj-tval.h"
 #include "obj-util.h"
 #include "player-attack.h"
