+++ src/obj-desc.c
@@ -20,8 +20,8 @@
 #include "obj-chest.h"
 #include "obj-desc.h"
 #include "obj-gear.h"
-#include "obj-identify.h"
 #include "obj-ignore.h"
+#include "obj-knowledge.h"
 #include "obj-tval.h"
 #include "obj-util.h"
