+++ src/obj-list.c
@@ -19,8 +19,8 @@
 #include "angband.h"
 #include "game-world.h"
 #include "obj-desc.h"
-#include "obj-identify.h"
 #include "obj-ignore.h"
+#include "obj-knowledge.h"
 #include "obj-list.h"
 #include "obj-pile.h"
 #include "obj-tval.h"
