+++ src/obj-pile.c
@@ -28,7 +28,6 @@
 #include "monster.h"
 #include "obj-desc.h"
 #include "obj-gear.h"
-#include "obj-identify.h"
 #include "obj-ignore.h"
 #include "obj-info.h"
 #include "obj-knowledge.h"
