+++ src/obj-slays.c
@@ -21,7 +21,6 @@
 #include "init.h"
 #include "mon-lore.h"
 #include "obj-desc.h"
-#include "obj-identify.h"
 #include "obj-knowledge.h"
 #include "obj-slays.h"
 #include "obj-util.h"
