+++ src/player-attack.c
@@ -32,7 +32,6 @@
 #include "monster.h"
 #include "obj-desc.h"
 #include "obj-gear.h"
-#include "obj-identify.h"
 #include "obj-knowledge.h"
 #include "obj-pile.h"
 #include "obj-slays.h"
