+++ src/player-calcs.c
@@ -27,7 +27,6 @@
 #include "mon-msg.h"
 #include "mon-util.h"
 #include "obj-gear.h"
-#include "obj-identify.h"
 #include "obj-ignore.h"
 #include "obj-knowledge.h"
 #include "obj-power.h"
@@ -1055,9 +1054,8 @@ void calc_inventory(struct player_upkeep *upkeep, struct object *gear,
 						upkeep->quiver[i] = current;
 						upkeep->quiver_cnt += current->number;
 
-						/* Notice stuff if it's first time in the quiver */
-						if (!object_was_worn(current))
-							object_learn_on_wield(player, current);
+						/* In the quiver counts as worn */
+						object_learn_on_wield(player, current);
 
 						/* Done with this slot */
 						break;
@@ -1101,9 +1099,8 @@ void calc_inventory(struct player_upkeep *upkeep, struct object *gear,
 		upkeep->quiver[i] = first;
 		upkeep->quiver_cnt += first->number;
 
-		/* Notice stuff if it's first time in the quiver */
-		if (!object_was_worn(first))
-			object_learn_on_wield(player, first);
+		/* In the quiver counts as worn */
+		object_learn_on_wield(player, first);
 	}
 
 	/* Note reordering */
