+++ src/project-player.c
@@ -22,7 +22,6 @@
 #include "init.h"
 #include "mon-desc.h"
 #include "obj-gear.h"
-#include "obj-identify.h"
 #include "obj-knowledge.h"
 #include "player-calcs.h"
 #include "player-timed.h"
