+++ src/ui-context.c
@@ -27,7 +27,6 @@
 #include "obj-chest.h"
 #include "obj-desc.h"
 #include "obj-gear.h"
-#include "obj-identify.h"
 #include "obj-ignore.h"
 #include "obj-info.h"
 #include "obj-tval.h"
