+++ src/ui-death.c
@@ -21,7 +21,6 @@
 #include "game-input.h"
 #include "init.h"
 #include "obj-desc.h"
-#include "obj-identify.h"
 #include "obj-info.h"
 #include "savefile.h"
 #include "store.h"
