+++ src/ui-knowledge.c
@@ -27,8 +27,8 @@
 #include "mon-util.h"
 #include "monster.h"
 #include "obj-desc.h"
-#include "obj-identify.h"
 #include "obj-ignore.h"
+#include "obj-knowledge.h"
 #include "obj-info.h"
 #include "obj-make.h"
 #include "obj-pile.h"
