+++ src/ui-player.c
@@ -21,8 +21,8 @@
 #include "init.h"
 #include "obj-desc.h"
 #include "obj-gear.h"
-#include "obj-identify.h"
 #include "obj-info.h"
+#include "obj-knowledge.h"
 #include "obj-util.h"
 #include "player.h"
 #include "player-calcs.h"
