+++ src/wiz-debug.c
@@ -29,7 +29,6 @@
 #include "monster.h"
 #include "obj-desc.h"
 #include "obj-gear.h"
-#include "obj-identify.h"
 #include "obj-make.h"
 #include "obj-pile.h"
 #include "obj-power.h"
